backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
google-genai==1.5.0
langgraph==0.4.1
langchain-core==0.3.51
httpx>=0.28.1
pytest==8.3.3
